CRUD comments

+++ back/controllers/topic.js
@@ -1,8 +1,10 @@
 const db = require("../models");
 const Topic = db.topics;
+const Comment = db.comments;
 const Op = db.Sequelize.Op;
 const fs = require('fs');
 const jwt = require('jsonwebtoken');
+const { error } = require("console");
 
 exports.createTopic = (req, res) => {
     const topic = {
@@ -30,29 +32,53 @@ exports.getAllTopics = (req, res) => {
     Topic.findAll({ where: condition })
       .then(data => {res.send(data);
     })
-      .catch(err => {
+      .catch(error => {
         res.status(500).send({ error
         });
       });
   };
 
 exports.modifyTopic = (req, res) => {
+    const token = req.headers.authorization.split(' ')[1];
+    //Décode le token
+    const decodedToken = jwt.verify(token, 'RANDOM_TOKEN_SECRET');
+    //Extrait l'id utilisateur et compare à celui extrait du token
+    const userId = decodedToken.userId;
+
     const id = req.params.id;
 
     Topic.findByPk(id)
-        .then(() => {
-            Topic.update(req.body, { where: { id: id } } )
-                .then(() => res.status(200).json({ message: 'Topic modifié !'}))
-                .catch(() => res.status(400).json({ error }));
+        .then(topic => {
+            if (topic.userId == userId) {
+                Topic.update(req.body, { where: { id: id } } )
+                    .then(() => res.status(200).json({ message: 'Topic modifié!'}))
+                    .catch(( error ) => res.status(400).json({ error }));
+            } else {
+                res.status(401).json({ message: 'Opération non autorisée' })
+            }
         })
         .catch(() => res.status(500).json({ error }));
 };
 
 exports.deleteTopic = (req, res) => {
+    const token = req.headers.authorization.split(' ')[1];
+    //Décode le token
+    const decodedToken = jwt.verify(token, 'RANDOM_TOKEN_SECRET');
+    //Extrait l'id utilisateur et compare à celui extrait du token
+    const userId = decodedToken.userId;
+
     const id = req.params.id;
 
-    Topic.destroy( { where: { id: id } } )
-        .then(() => res.status(200).json({ message: 'Topic supprimé !'}))
-        .catch(error => res.status(400).json({ error }));
+    Topic.findByPk(id)
+        .then(topic => {
+            if (topic.userId == userId) {
+                Topic.destroy( { where: { id: id } } )
+                    .then(() => res.status(200).json({ message: 'Topic modifié!'}))
+                    .catch(( error ) => res.status(400).json({ error }));
+            } else {
+                res.status(401).json({ message: 'Opération non autorisée' })
+            }
+        })
+        .catch((error) => res.status(500).json({ error }));
 };
